requirements.txt
Django==1.7.1
Fabric==1.8.1
argparse==1.2.1
django-appconf==0.6
django-bootstrap3==4.11.0
django-braces==1.4.0
django-compressor==1.4
django-debug-toolbar==1.2.1
django-extensions==1.4.0
django-grappelli==2.6.1
django-redis==3.8.0
djrill==1.2.0
ecdsa==0.11
gunicorn==18.0
jsonschema==2.4.0
paramiko==1.15.1
psycopg2==2.5.2
pycrypto==2.6.1
pytz==2014.4
redis==2.10.3
requests==2.5.1
six==1.8.0
sqlparse==0.1.14
wsgiref==0.1.2
